Change conc view vectors, remove some unused code
The view vectors in FF were made so the origin is in the middle of the hull, but that is not needed, so we can just have the origin at the bottom of the hull as usual. This should also fix the crouch stuck bug.

Also changed the conc throw vector to accomodate for this change.